Started the iterating over Data at the startIndex.
Data uses slices which may not start at 0. When a Data is supplied which does not start at 0, the first index previously used would throw a bad access exception.

This happened for example with a Data value returns from CBOR previously. It does not per-se start with a slice at 0.